Instrumentation of vtable debug metadata.
We previously inferred vtable names from LLVM IR but ran into a number of limitations
due to the lack of information in the IR. We now do it from within clang instead.
A command-line option -gsoaap has been added that adds to each virtual call, two pieces
of metadata:

1) vtable name/var for the class defining the function being called
2) vtable name/var for the class corresponding to the static type of the receiver

The former is necessary so that we can find where the first callee is in the static type (by using
object offsets) and the second is necessary so that we only infer callees from the static
type downwards (as there may be other subclasses of the defining class whose implementations
are irrelevant.