Fix bug in clang++ -gsoaap leading to assertion failures in clang
Previously we were manually constructing mangled vtable names using MangleContext::mangleName(...), however the behaviour of this appears to have changed and is causing assertion failures when using a debug build of clang.

The implementation now obtains the mangled name from the VTable global variable. We were previously obtaining this global variable to force the vtable to be generated anyway.