Fix -gsoaap (again)
Previously we were obtaining a vtable global variable to obtain the vtable's
mangled name. This caused vtables to be generated earlier than expected leading
to assertion failures. We now use ItaniumMangleContext to get the mangled named
and only obtain the vtable variable in the case of classes within anonymous
namespaces.