Pass output parameter to obtain CallInst emitted for a virtual call.
Previously, after IR had been emitted for a virtual call, we were using the
IRBuilder to obtain the CallInst* by iterating backwards until the first
CallInst* was found. Sometimes llvm will insert additional CallInst
instructions such as when a function returns a temporary struct that is then
copied to a local variable using llvm.memcpy. Thus, the most reliable way to
obtain the *actual* callinst for the member call is to pass in an output
parameter that is assigned to once it has been emitted.